Feat: add editor camera detection to map report
I attempted to be clever and make a little lambda inline since this Yes/No thing is only used here atm.
Sadly it didn't appear to work, so I've moved it out as a constexpr lambda.
I believe I was allowed to make it constexpr (not consteval though) because it returns two literals, which therefore are known at compile-time as you'd expect.
Interesting stuff.